+++ src/examples/GraphTheory/Grids/GridDFSPath.stories.tsx
@@ -1,11 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/react";
 import clsx from "clsx";
-
-import { useEffect, useState } from "react";
-
-import type { Coordinates } from ".";
-import { Grid, dfsPath as findDfsPath, generateTestMatrix } from ".";
-
+import { Grid } from ".";
 import { GraphKey } from "../GraphKey";
 import {
   BG_PLAYER_PATH_COLOR,
@@ -14,10 +9,8 @@ import {
   BG_TARGET_COLOR,
   BG_TERRAIN_COLOR,
   BG_VISITED_TILE_COLOR,
-  PLAYER_START_FILL_COLOR,
-  TARGET_FILL_COLOR,
 } from "./constants";
-import type { GridMatrix } from "./GridTypes";
+import useGridState from "./hooks/useGridState";
 
 const meta: Meta<typeof Grid> = {
   component: Grid,
@@ -27,157 +20,12 @@ const meta: Meta<typeof Grid> = {
   },
   decorators: [
     (Story) => {
-      const [originalMatrix, setOriginalMatrix] = useState<GridMatrix | null>(
-        null,
-      );
-      const [gridVisualisationMatrix, setGridVisualisationMatrix] =
-        useState<GridMatrix | null>(null);
-
-      const [nextClickTileType, setNextClickTileType] = useState<
-        "start" | "target"
-      >("start");
-      const [startCoordinates, setStartCoordinates] =
-        useState<Coordinates | null>(null);
-      const [targetCoordinates, setTargetCoordinates] =
-        useState<Coordinates | null>(null);
-
-      const [tileColorOverride, setTileColorOverride] = useState({});
-      const [path, setPath] = useState<Coordinates[] | null>([]);
-      const [isRunning, setIsRunning] = useState(false);
-
-      const pickPathTile = (tile: Coordinates) => {
-        // If wall tile you can't pick it
-        if (
-          !gridVisualisationMatrix ||
-          gridVisualisationMatrix[tile.y][tile.x] === 1
-        )
-          return;
-
-        if (nextClickTileType === "start") {
-          setTargetCoordinates(() => null);
-          setStartCoordinates(() => tile);
-          // toggle so next click sets target
-          setNextClickTileType(() => "target");
-        }
-
-        if (nextClickTileType === "target") {
-          setTargetCoordinates(() => tile);
-          // toggle so next click sets start
-          setNextClickTileType(() => "start");
-        }
-      };
-
-      useEffect(() => {
-        const newMatrix = generateTestMatrix({
-          placePlayer: false,
-          placeTargetTile: false,
-        });
-
-        setOriginalMatrix(() => newMatrix);
-        setGridVisualisationMatrix(() => structuredClone(newMatrix));
-      }, []);
-
-      // Handle state updates
-      // Handle setting the start/ target coordinates after user clicks
-      useEffect(() => {
-        if ((!startCoordinates && !targetCoordinates) || isRunning) return;
-
-        // Target tile was clicked
-        if (nextClickTileType === "start" && targetCoordinates) {
-          setTileColorOverride((prev) => ({
-            ...prev,
-            targetTile: {
-              ...targetCoordinates,
-              color: TARGET_FILL_COLOR,
-            },
-          }));
-        }
-
-        if (nextClickTileType === "target" && startCoordinates) {
-          // Start tile was clicked
-          // Reset override colors colors
-          setTileColorOverride(() => ({}));
-
-          // Reset to original UI colors to start picking tiles again
-          setGridVisualisationMatrix(() => structuredClone(originalMatrix));
-
-          setTileColorOverride((prev) => ({
-            ...prev,
-            startTile: {
-              ...startCoordinates,
-              color: PLAYER_START_FILL_COLOR,
-            },
-          }));
-        }
-      }, [startCoordinates, targetCoordinates, nextClickTileType]);
-
-      useEffect(() => {
-        if (
-          startCoordinates &&
-          targetCoordinates &&
-          !isRunning &&
-          originalMatrix
-        ) {
-          setIsRunning(() => true);
-          findDfsPath({
-            grid: originalMatrix,
-            startCoordinates,
-            targetCoordinates,
-            setGridVisualisationMatrix,
-            stepInterval: 200,
-          })
-            .then((path) => {
-              if (path) console.log("✅ Path found:", path);
-
-              setPath(() => path);
-              setIsRunning(false);
-            })
-            .catch(() => {
-              setIsRunning(false);
-            });
-        }
-      }, [startCoordinates, targetCoordinates]);
-
-      type UpdateGridVisualisationMatrix = (args: {
-        x: number;
-        y: number;
-        tileIdentifier?: number;
-      }) => void;
-
-      const updateGridVisualisationMatrix: UpdateGridVisualisationMatrix = ({
-        x,
-        y,
-        tileIdentifier /* 0 - 5 */,
-      }) => {
-        setGridVisualisationMatrix((prevMatrix) => {
-          if (prevMatrix === null) return prevMatrix;
-          const newMatrix = prevMatrix.map((row) => row.slice());
-
-          // Don't allow pathing on wall
-          if (newMatrix[y][x] === 1) {
-            return prevMatrix;
-          }
-
-          // 5 = will be used to demonstrate all processed tiles as the alg steps still WIP though
-          // 0 = reset
-          newMatrix[y][x] =
-            tileIdentifier || tileIdentifier === 0 ? tileIdentifier : 5;
-          return newMatrix;
-        });
-      };
-
-      // Draw path
-      useEffect(() => {
-        if (!path || !path.length || nextClickTileType === "target") return;
-        // Tiles from start to target
-        path.forEach(({ x, y }) => {
-          updateGridVisualisationMatrix({ x, y, tileIdentifier: 4 });
-        });
-      }, [path]);
-
-      // useEffect(() => {
-      //   console.log("gridVisualisationMatrix", gridVisualisationMatrix);
-      // }, [gridVisualisationMatrix]);
+      const {
+        originalMatrix,
+        gridVisualisationMatrix,
+        tileColorOverride,
+        pickPathTile,
+      } = useGridState("dfs");
 
       return (
         <div
